### 2017-08-24 #375 @whedon
[PRE REVIEW]: bikedata
**Submitting author:** @mpadge (<a href="http://orcid.org/0000-0003-2172-5265">Mark Padgham</a>)
**Repository:** <a href="https://github.com/ropensci/bikedata" target ="_blank">https://github.com/ropensci/bikedata</a>
**Version:** v0.0.4
