### 2017-08-25 #375 @maelle
(not interfering with the previous comment, just for info) This submission has been accepted to rOpenSci. The review thread can be found at https://github.com/ropensci/onboarding/issues/116
