### 2017-08-25 #375 @mpadge
@arfon Fair question on something I didn't directly address in the current [`paper.md`](https://github.com/ropensci/bikedata/blob/master/paper.md). It has very direct research application through opening up - curating, one might even say - arguably the worlds largest open ongoing dataset on human mobility. Without the package, these kinds of data can only be accessed in a very ad-hoc and piecemeal fashion. Moreover, there is quite arguably no other equivalent data set on human mobility that is both open and constantly evolving and updating.

Shall I update the `paper.md` to incorporate some of those points?
