### 2017-08-25 #375 @karthik
@mpadge Yes please, the more you can articulate a direct research application, the more likely it is to get approved here at JOSS.
