### 2017-08-25 e538b48
tweak paper.md
