### 2017-09-28 0f48df7
ah, that old stringAsFactors again ...
