### 2017-09-29 5171bb6
fix vignette link (mpadge -> ropensci)
