+++ README.Rmd
@@ -52,7 +52,7 @@ of cyclists.
 
 The following provides a brief overview of package functionality. For more
 detail, see the
-[vignette](https://mpadge.github.io/bikedata/articles/bikedata.html).
+[vignette](https://ropensci.github.io/bikedata/articles/bikedata.html).
 
 ------
 
