### 2017-10-02 cf4a326
version bump 0.0.4 -> 0.0.4.99
