+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: bikedata
 Title: Download and Aggregate Data from Public Hire Bicycle Systems
-Version: 0.0.4
+Version: 0.0.4.99
 Authors@R: c(
     person("Mark", "Padgham", email = "mark.padgham@email.com", role = c("aut", "cre")),
     person("Richard", "Ellison", role = "aut"),
