### 2017-10-02 ae749a0
store member/non-member for NABSA systems
