+++ R/tripmat.R
@@ -315,7 +315,7 @@ bike_tripmat <- function (bikedb, city, start_date, end_date,
     if (!missing (weekday))
         x <- c (x, 'weekday' = list (convert_weekday (weekday)))
 
-    if ( !missing (birth_year) | !missing (gender) &
+    if ( (!missing (birth_year) | !missing (gender)) &
         !city %in% (c ('bo', 'ch', 'ny')))
         stop ('Only Boston, Chicago, and New York provide demographic data')
     if ( !missing (member) & !city %in% c ('bo', 'ch', 'ny', 'la', 'ph'))
