### 2017-10-09 27e2e13
add bike_distmat fn
also create check_city_arg fn
