+++ tests/testthat/test-db-stats.R
@@ -6,7 +6,7 @@ bikedb <- system.file ('db', 'testdb.sqlite', package = 'bikedata')
 
 test_that ('can not read all data', {
                expect_error (trips <- bike_tripmat (bikedb),
-                             'Calls to bike_tripmat must specify city')
+                         'bikedb contains multiple cities; please specify one')
                expect_error (trips <- bike_tripmat (bikedb, city = 'aa'),
                              'city not recognised')
 })
