+++ DESCRIPTION
@@ -20,6 +20,7 @@ Depends:
     R (>= 3.0)
 Imports:
     DBI,
+    dodgr,
     httr,
     lubridate,
     magrittr,
@@ -38,6 +39,8 @@ Suggests:
 LinkingTo: 
     BH,
     Rcpp
+Remotes:
+    gmost/dodgr
 VignetteBuilder: knitr
 SystemRequirements: C++11
 NeedsCompilation: yes
