+++ NAMESPACE
@@ -4,6 +4,7 @@ export(bike_daily_trips)
 export(bike_datelimits)
 export(bike_db_totals)
 export(bike_demographic_data)
+export(bike_distmat)
 export(bike_latest_files)
 export(bike_rm_db)
 export(bike_rm_test_data)
@@ -22,6 +23,7 @@ importFrom(DBI,dbGetQuery)
 importFrom(DBI,dbSendQuery)
 importFrom(RSQLite,SQLite)
 importFrom(Rcpp,evalCpp)
+importFrom(dodgr,dodgr_dists)
 importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(lubridate,ddays)
@@ -30,6 +32,7 @@ importFrom(lubridate,ymd)
 importFrom(magrittr,"%>%")
 importFrom(methods,as)
 importFrom(reshape2,dcast)
+importFrom(reshape2,melt)
 importFrom(tibble,as_tibble)
 importFrom(tibble,tibble)
 importFrom(utils,data)
