### 2017-10-09 b81fbb7
update news, travis.yml
