### 2017-10-09 5fb9530
finish bike_distmat fn
