+++ man/bike_distmat.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dists.R
+% Please edit documentation in R/distmat.R
 \name{bike_distmat}
 \alias{bike_distmat}
 \title{Extract station-to-station distance matrix}
@@ -24,8 +24,8 @@ stn-to, distance) is returned.}
 }
 \value{
 If \code{long = FALSE}, a square matrix of numbers of trips between
-each station, otherwise a long-form data.frame with three columns of of
-(start_station, end_station, num_trips)
+each station, otherwise a long-form \pkg{tibble} with three columns of of
+(start_station_id, end_station_id, distance)
 }
 \description{
 Extract station-to-station distance matrix
