+++ NAMESPACE
@@ -14,6 +14,7 @@ export(bike_tripmat)
 export(bike_write_test_data)
 export(dl_bikedata)
 export(index_bikedata_db)
+export(match_matrices)
 export(store_bikedata)
 importFrom(DBI,dbBind)
 importFrom(DBI,dbClearResult)
