### 2017-10-09 902ba03
rm osx devel from travis.yml
