+++ .travis.yml
@@ -15,12 +15,12 @@ matrix:
         - os: linux
           r: devel
           dist: trusty
-        - os: osx # currently lacks dbplyr - reactivate soon (12Jun17)?
+        - os: osx
           r: release
           #before_install: ''
           #disable_homebrew: true
-        - os: osx
-          r: devel
+          #- os: osx
+          #r: devel
           #before_install: ''
           #disable_homebrew: true
 
