### 2017-10-12 0e7fb17
alias dl_bikedata to download_bikedata
