+++ man/dl_bikedata.Rd
@@ -2,9 +2,12 @@
 % Please edit documentation in R/dl-bikedata.R
 \name{dl_bikedata}
 \alias{dl_bikedata}
+\alias{download_bikedata}
 \title{Download hire bicycle data}
 \usage{
 dl_bikedata(city, data_dir = tempdir(), dates = NULL, quiet = FALSE)
+
+download_bikedata(city, data_dir = tempdir(), dates = NULL, quiet = FALSE)
 }
 \arguments{
 \item{city}{City for which to download bike data, or name of corresponding
