+++ NAMESPACE
@@ -13,6 +13,7 @@ export(bike_summary_stats)
 export(bike_tripmat)
 export(bike_write_test_data)
 export(dl_bikedata)
+export(download_bikedata)
 export(index_bikedata_db)
 export(match_matrices)
 export(store_bikedata)
