### 2017-10-12 7ed509b
fix convert_dates_to_filenames
