### 2017-10-12 28c77ea
only return bike_stations with non-zero lon-lats
