### 2017-10-12 6187499
expand *nix-style ~ directory args to home
