### 2017-10-12 4ac3494
ensure store_bikedata only adds new data
previous code was buggy, and re-added data atop already existing data
