### 2017-10-12 c25fffb
fix dates arg for store_bikedata
