### 2017-10-12 0747b19
fix repeat store_bikedata for chicago
