+++ R/store-bikedata.R
@@ -118,7 +118,7 @@ store_bikedata <- function (bikedb, city, data_dir, dates = NULL, quiet = FALSE)
                 message ('Unzipping raw data files for ', ci, ' ...')
         }
         if (ci == 'ch')
-            flists <- bike_unzip_files_chicago (data_dir, bikedb)
+            flists <- bike_unzip_files_chicago (data_dir, bikedb, dates)
         else
             flists <- bike_unzip_files (data_dir, bikedb, ci, dates)
 
@@ -142,15 +142,16 @@ store_bikedata <- function (bikedb, city, data_dir, dates = NULL, quiet = FALSE)
             if (ci == 'ch')
             {
                 ch_stns <- bike_get_chicago_stations (flists)
-                nstations <- rcpp_import_stn_df (bikedb, ch_stns, 'ch') #nolint
+                if (nrow (ch_stns) > 0)
+                    nstations <- rcpp_import_stn_df (bikedb, ch_stns, 'ch')
             } else if (ci == 'lo')
             {
                 lo_stns <- bike_get_london_stations ()
-                nstations <- rcpp_import_stn_df (bikedb, lo_stns, 'lo') #nolint
+                nstations <- rcpp_import_stn_df (bikedb, lo_stns, 'lo')
             } else if (ci == 'dc')
             {
                 dc_stns <- bike_get_dc_stations ()
-                nstations <- rcpp_import_stn_df (bikedb, dc_stns, 'dc') #nolint
+                nstations <- rcpp_import_stn_df (bikedb, dc_stns, 'dc')
             }
 
             # main step: Import trips
@@ -366,7 +367,6 @@ get_flist_city <- function (data_dir, bikedb, city)
 #' If no directory specified, it is presumed to be in \code{tempdir()}.
 #' @param city City for which files are to be added to database
 #' @param dates If specified, data will be unzipped only for these dates
-#' specified
 #'
 #' @return List of three vectors of file names:
 #' \itemize{
@@ -387,7 +387,7 @@ bike_unzip_files <- function (data_dir, bikedb, city, dates)
                                  city = city)
     # These are only those files in data_dir but **not** in bikedb. Remove any
     # not in dates:
-    if (!missing (dates))
+    if (!is.null (dates))
     {
         dates <- bike_convert_dates (dates) %>%
             expand_dates_to_range %>%
@@ -435,6 +435,7 @@ bike_unzip_files <- function (data_dir, bikedb, city, dates)
 #' @param data_dir Directory containing data files
 #' @param bikedb A string containing the path to the SQLite3 database 
 #' If no directory specified, it is presumed to be in \code{tempdir()}.
+#' @param dates If specified, data will be unzipped only for these dates
 #'
 #' @return List of three vectors of file names:
 #' \itemize{
@@ -453,10 +454,19 @@ bike_unzip_files <- function (data_dir, bikedb, city, dates)
 #' file.
 #'
 #' @noRd
-bike_unzip_files_chicago <- function (data_dir, bikedb)
+bike_unzip_files_chicago <- function (data_dir, bikedb, dates)
 {
     flist_zip <- get_flist_city (data_dir = data_dir, bikedb = bikedb,
                                  city = 'ch')
+    if (!is.null (dates))
+    {
+        dates <- bike_convert_dates (dates) %>%
+            expand_dates_to_range %>%
+            convert_dates_to_filenames (city = city)
+        indx <- which (grepl (paste (dates, collapse = "|"), flist_zip))
+        flist_zip <- flist_zip [indx]
+    }
+
     flist_zip <- get_new_datafiles (bikedb, flist_zip)
     existing_csv_files <- list.files (data_dir, pattern = "Divvy.*\\.csv")
     if (length (existing_csv_files) == 0)
