### 2017-10-12 90f5c1a
fix bike_summary_stats
