### 2017-10-12 cfe805b
fix up datetime storage for NABSA systems
they are actually NOT in consistent formats! This now handles both formats used thus far
