### 2017-10-12 f787517
change bike_tripmat stadardise default F -> T
