### 2017-10-13 dbc877e
revert bike_tripmat standardise default T -> F
