### 2017-10-13 64c950c
fix get_new_datafiles fn
