### 2017-10-13 5d61516
fix convert_dates_to_filenames for London
