+++ R/date-time-functions.R
@@ -235,8 +235,13 @@ convert_dates_to_filenames <- function (x, city = 'ny')
             x <- c ('2013', x)
     } else if (city == 'lo')
     {
+        indx1 <- which (yy < 2015)
+        indx2 <- which (yy >= 2015)
+        x1 <- yy [indx1]
+        x <- x [indx2]
         mm <- month.abb [as.numeric (substring (x, 5, 6))]
-        x <- paste0 (mm, yy)
+        x <- c (paste0 (mm, yy), paste0 (mm, substring (yy, 3, 4)))
+        x <- unique (c (x, x1))
     } else if (city %in% c ('dc', 'la', 'ph'))
     {
         # LA uses both "YYYY_QX" and "QX_YYYY"
