### 2017-10-13 b61d1a8
deal with junk nyc zip files
