### 2017-10-13 d805a4a
switch off problematic cran windows test
