### 2017-10-17 16fc8c9
junk nyc files are gone now - must have been a glitch?
