### 2017-10-17 6f23094
improves dealing with junk AWS files
