### 2017-10-18 ba50d18
fix reading most recent NY file formats
bloody well changed format once again
