### 2017-10-18 fac6895
ensure ny stations exclude tests with lat/lon=0
