### 2017-10-19 4bf561e
add rm_xy_outliers fn for distmat
