### 2017-10-19 d9567d1
fix distmat
