+++ man/bike_match_matrices.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/distmat.R
-\name{match_matrices}
-\alias{match_matrices}
+\name{bike_match_matrices}
+\alias{bike_match_matrices}
 \title{Match rows and columns of distance and trip matrices}
 \usage{
-match_matrices(mat1, mat2)
+bike_match_matrices(mat1, mat2)
 }
 \arguments{
 \item{mat1}{A wide- or long-form trip or distance matrix returned from
