+++ NAMESPACE
@@ -6,6 +6,7 @@ export(bike_db_totals)
 export(bike_demographic_data)
 export(bike_distmat)
 export(bike_latest_files)
+export(bike_match_matrices)
 export(bike_rm_db)
 export(bike_rm_test_data)
 export(bike_stations)
@@ -15,7 +16,6 @@ export(bike_write_test_data)
 export(dl_bikedata)
 export(download_bikedata)
 export(index_bikedata_db)
-export(match_matrices)
 export(store_bikedata)
 importFrom(DBI,dbBind)
 importFrom(DBI,dbClearResult)
