+++ R/RcppExports.R
@@ -147,86 +147,3 @@ rcpp_create_city_index <- function(bikedb, reindex) {
     .Call(`_bikedata_rcpp_create_city_index`, bikedb, reindex)
 }
 
-#' strtokm
-#'
-#' A string delimiter function based on strtok
-#' Accessed from StackOverflow (using M Oehm):
-#' http://stackoverflow.com/questions/29847915/implementing-strtok-whose-delimiter-has-more-than-one-character
-#'
-#' @noRd
-NULL
-
-#' str_token
-#'
-#' A delimiter function for comma-separated std::string
-#'
-#' @param line The line of text to be tokenised
-#' @param delim The desired delimiter
-#'
-#' @return Next token
-#'
-#' @noRd
-NULL
-
-#' rm_dos_end
-#'
-#' Remove dos line ending from a character string
-#'
-#' @noRd
-NULL
-
-#' line_has_quotes
-#'
-#' Determine whether or not fields within a line are separated by double quotes
-#' and a comma, or just comma separated.
-#'
-#' @param line Character string with or without double-quote-comma separation
-#'
-#' @return true if line is delimited by double quotes and commas, false if
-#' commas only.
-#'
-#' @noRd
-NULL
-
-#' convert_datetime_ny
-#'
-#' Datetime strings for NYC change between 08/2014 and 09/2014 from
-#' yyyy-mm-dd HH:MM:SS to either m/d/yyyy HH:MM:SS or m/d/yyyy H:M
-#'
-#' @noRd
-NULL
-
-#' convert_datetime_ch
-#'
-#' Datetime strings for Chicago vary between the following formats:
-#' YYYY-MM-DD hh:mm
-#' M/D/YYYY hh:mm
-#' M/D/YYYY hh:mm:ss
-#'
-#' @noRd
-NULL
-
-#' convert_datetime_nabsa
-#'
-#' North American Bike Share Association (LA and Philadelphia) have identical
-#' formats, but they still change from M/D/YYYY H:MM to a more regular
-#' YYYY-MM-DD HH:MM:SS.
-#'
-#' @noRd
-NULL
-
-#' convert_datetime_dc
-#'
-#' Datetime strings for DC are either M/D/YYYY h:mm, where "mm" is always
-#' 0-padded, or YYYY-MM-DD hh:mm, where "hh" is always 0-padded
-#'
-#' @noRd
-NULL
-
-#' convert_datetime_lo
-#'
-#' Only issue with London is sometimes seconds are present; sometimes not
-#'
-#' @noRd
-NULL
-
