+++ src/sqlite3db-utils.h
@@ -14,6 +14,7 @@
 
 #define BUFFER_SIZE 512
 
+int get_max_trip_id (sqlite3 * dbcon);
 int get_max_stn_id (sqlite3 * dbcon);
 int get_stn_table_size (sqlite3 * dbcon);
 
@@ -29,7 +30,7 @@ int get_max_trip_id (sqlite3 * dbcon)
     sqlite3_stmt * stmt;
     char qry_id [BUFFER_SIZE];
     int rc = sprintf(qry_id, "SELECT MAX(id) FROM trips");
-    rc = sqlite3_prepare_v2(dbcon, qry_id, BUFFER_SIZE, &stmt, NULL);
+    rc = sqlite3_prepare_v2(dbcon, qry_id, BUFFER_SIZE, &stmt, nullptr);
     rc = sqlite3_step (stmt);
     int max_trip_id = sqlite3_column_int (stmt, 0);
     sqlite3_reset (stmt);
@@ -50,7 +51,7 @@ int get_max_stn_id (sqlite3 * dbcon)
     sqlite3_stmt * stmt;
     char qry_id [BUFFER_SIZE];
     sprintf(qry_id, "SELECT MAX(id) FROM stations");
-    int rc = sqlite3_prepare_v2(dbcon, qry_id, BUFFER_SIZE, &stmt, NULL);
+    int rc = sqlite3_prepare_v2(dbcon, qry_id, BUFFER_SIZE, &stmt, nullptr);
     rc = sqlite3_step (stmt);
     int max_stn_id = sqlite3_column_int (stmt, 0);
     sqlite3_reset (stmt);
@@ -71,7 +72,7 @@ int get_stn_table_size (sqlite3 * dbcon)
     sqlite3_stmt * stmt;
     char qry_id [BUFFER_SIZE];
     sprintf(qry_id, "SELECT COUNT(*) FROM stations");
-    int rc = sqlite3_prepare_v2(dbcon, qry_id, BUFFER_SIZE, &stmt, NULL);
+    int rc = sqlite3_prepare_v2(dbcon, qry_id, BUFFER_SIZE, &stmt, nullptr);
     rc = sqlite3_step (stmt);
     int num_stns = sqlite3_column_int (stmt, 0);
     sqlite3_reset (stmt);
