+++ src/read-city-files.h
@@ -18,7 +18,7 @@
 
 // NOTE: Return values are only used for LA
 
-unsigned read_one_line_nyc (sqlite3_stmt * stmt, char * line,
+unsigned int read_one_line_nyc (sqlite3_stmt * stmt, char * line,
         std::map <std::string, std::string> * stationqry, const char * delim);
 unsigned int read_one_line_nyc_standard (sqlite3_stmt * stmt,
         std::string &in_line2,
@@ -27,19 +27,19 @@ unsigned int read_one_line_nyc_mixed (sqlite3_stmt * stmt,
         std::string &in_line2,
         std::map <std::string, std::string> * stationqry);
 
-unsigned read_one_line_boston (sqlite3_stmt * stmt, char * line,
+unsigned int read_one_line_boston (sqlite3_stmt * stmt, char * line,
         std::map <std::string, std::string> * stationqry);
-unsigned read_one_line_chicago (sqlite3_stmt * stmt, char * line,
+unsigned int read_one_line_chicago (sqlite3_stmt * stmt, char * line,
         const char * delim);
-unsigned read_one_line_dc (sqlite3_stmt * stmt, char * line, 
+unsigned int read_one_line_dc (sqlite3_stmt * stmt, char * line, 
         std::map <std::string, std::string> &stn_map, 
         std::unordered_set <std::string> &stn_ids,
         bool id, bool end_date_first);
 std::string convert_dc_stn_name (std::string &station_name, bool id,
         std::map <std::string, std::string> &stn_map);
-unsigned read_one_line_london (sqlite3_stmt * stmt, char * line);
+unsigned int read_one_line_london (sqlite3_stmt * stmt, char * line);
 std::string add_0_to_time (std::string time);
-unsigned read_one_line_nabsa (sqlite3_stmt * stmt, char * line,
+unsigned int read_one_line_nabsa (sqlite3_stmt * stmt, char * line,
         std::map <std::string, std::string> * stationqry,
         std::string city);
 
@@ -87,19 +87,19 @@ unsigned int read_one_line_nyc_standard (sqlite3_stmt * stmt,
         // Example of the following on L#19 of 2014-07
         boost::replace_all(in_line2, "\\N","\"\"");
         duration = strtokm (&in_line2[0u], "\""); //First double speech marks
-        duration = strtokm (NULL, delim); 
+        duration = strtokm (nullptr, delim); 
     } else
         duration = strtokm (&in_line2[0u], delim);
 
 
-    std::string start_time = convert_datetime_ny (strtokm (NULL, delim)); 
-    std::string end_time = convert_datetime_ny (strtokm (NULL, delim)); 
-    std::string start_station_id = strtokm (NULL, delim);
+    std::string start_time = convert_datetime_ny (strtokm (nullptr, delim)); 
+    std::string end_time = convert_datetime_ny (strtokm (nullptr, delim)); 
+    std::string start_station_id = strtokm (nullptr, delim);
     start_station_id = "ny" + start_station_id;
     if (stationqry->count(start_station_id) == 0) {
-        std::string start_station_name = strtokm (NULL, delim);
-        std::string start_station_lat = strtokm (NULL, delim);
-        std::string start_station_lon = strtokm (NULL, delim);
+        std::string start_station_name = strtokm (nullptr, delim);
+        std::string start_station_lat = strtokm (nullptr, delim);
+        std::string start_station_lon = strtokm (nullptr, delim);
         if (start_station_lat != "0.0" && start_station_lon != "0.0")
         {
             (*stationqry)[start_station_id] = "(\'ny\',\'" + 
@@ -108,17 +108,17 @@ unsigned int read_one_line_nyc_standard (sqlite3_stmt * stmt,
         }
     }
     else {
[…]

[…]
