### 2017-11-20 1c921ae
import stats::sd
