+++ R/bikedata-package.R
@@ -63,6 +63,7 @@
 #' @importFrom Rcpp evalCpp
 #' @importFrom RSQLite SQLite
 #' @importFrom reshape2 dcast melt
+#' @importFrom stats sd
 #' @importFrom tibble as_tibble tibble
 #' @importFrom utils data menu read.csv tail type.convert unzip write.csv zip
 #' @importFrom xml2 xml_children xml_find_all
