+++ NAMESPACE
@@ -35,6 +35,7 @@ importFrom(magrittr,"%>%")
 importFrom(methods,as)
 importFrom(reshape2,dcast)
 importFrom(reshape2,melt)
+importFrom(stats,sd)
 importFrom(tibble,as_tibble)
 importFrom(tibble,tibble)
 importFrom(utils,data)
