### 2017-11-20 24a777a
upgrade sqlite3 v3.19 -> v3.21
