### 2017-11-20 65e98ef
rm remotes from descr
