### 2017-11-23 814c273
update cran-comments.md
