+++ cran-comments.md
@@ -8,15 +8,12 @@
 * Windows Visual Studio 2015 (on `appveyor`; `x64`): R-release, R-devel
 * win-builder (R-release, R-devel, R-oldrelease)
 
-Package also checked using both local memory sanitzer and `rocker/r-devel-san`
-with clean results.
+Package also checked using both local memory sanitzer and `rocker/r-devel-san` with clean results.
 
 
 # R CMD check results
 
 The following NOTEs are generated:
 1. Potential mis-spelling of "Los" and "Angeles" (on R-release, R-oldrelease).
-2. Authors@R field gives persons with no valid roles: "rev" (on R-release,
-   R-oldrelease) - These are accepted roles on R-devel.
-3. Note regarding package size due to necessity of internally bundling SQLite3
-   header library, along with extensive internal C++ routines.
+2. Authors@R field gives persons with no valid roles: "rev" (on R-release, R-oldrelease) - These are accepted roles on R-devel.
+3. Note regarding package size due to necessity of internally bundling SQLite3 header library, along with extensive internal C++ routines.
