### 2017-11-27 #471 @arfon
@karthik - could you link to the rOpenSci review here please?
