### 2017-11-27 #471 @arfon
@mpadge - our paper processing system expects there to be references associated with each paper. Is it possible for you to cite an related/existing work?
