### 2017-11-27 #471 @mpadge
@arfon yep, of course. I'll put in a `paper.bib` tomorrow (Tues 28/11). Great that this is progressing!
