### 2017-11-27 #471 @karthik
@arfon Here is the issue: https://github.com/ropensci/onboarding/issues/116
