+++ paper.bib
@@ -1,0 +1,49 @@
+@Article{Austwick2013,
+    Title                    = {The Structure of Spatial Networks and Communities in Bicycle Sharing Systems},
+    Author                   = {Austwick, Martin Zaltz AND O’Brien, Oliver AND Strano, Emanuele AND Viana, Matheus},
+    Journal                  = {PLoS ONE},
+    Year                     = {2013},
+    Month                    = {09},
+    Number                   = {9},
+    Pages                    = {e74685},
+    Volume                   = {8},
+    Doi                      = {10.1371/journal.pone.0074685}
+} 
+
+@Article{Borgnat2011,
+    Title                    = {Shared Bicycles in a City: A Signal Processing and Data Analysis Perspective},
+    Author                   = {P. Borgnat and C. Robardet and J.-B. Rouquier and E. Fleury and P. Abry and P. Flandrin},
+    Journal                  = {Advances in Complex Systems},
+    Year                     = {2011},
+    Pages                    = {415--438},
+    Volume                   = {14},
+    Doi                      = {10.1142/S0219525911002950}
+}
+
+@InCollection{Guenther2013,
+    Title                    = {Journey Data Based Arrival Forecasting for Bicycle Hire Schemes},
+    Author                   = {Guenther, MarcelC. and Bradley, JeremyT.},
+    Booktitle                = {Analytical and Stochastic Modeling Techniques and Applications},
+    Publisher                = {Springer Berlin Heidelberg},
+    Year                     = {2013},
+    Editor                   = {Dudin, Alexander and De Turck, Koen},
+    Pages                    = {214-231},
+    Series                   = {Lecture Notes in Computer Science},
+    Volume                   = {7984},
+    Doi                      = {10.1007/978-3-642-39408-9_16},
+    ISBN                     = {978-3-642-39407-2}
+}
+
+
+@Article{Padgham2012,
+    Title                    = {Human Movement Is Both Diffusive and Directed},
+    Author                   = {Padgham, , Mark},
+    Journal                  = {PLoS ONE},
+    Year                     = {2012},
+    Month                    = {05},
+    Number                   = {5},
+    Pages                    = {e37754},
+    Volume                   = {7},
+    Doi                      = {10.1371/journal.pone.0037754}
+}
+
