### 2017-11-28 #471 @mpadge
[`paper.bib`](https://github.com/ropensci/bikedata/blob/master/paper.bib) now inserted and referenced in [`paper.md`](https://github.com/ropensci/bikedata/blob/master/paper.md). Let me know if there are any other requirements.
