### 2017-11-30 2075ffa
add keywords to README.Rmd
