### 2017-11-30 #471 @arfon
@mpadge - please merge this PR which fixes up the YAML metadata in your paper: https://github.com/ropensci/bikedata/pull/50
