### 2017-12-01 b24102f
editing description and generated codemeta.json file
