+++ DESCRIPTION
@@ -2,7 +2,9 @@ Package: bikedata
 Title: Download and Aggregate Data from Public Hire Bicycle Systems
 Version: 0.1.0
 Authors@R: c(
-    person("Mark", "Padgham", email = "mark.padgham@email.com", role = c("aut", "cre")),
+    person("Mark", "Padgham", email = "mark.padgham@email.com",
+        role = c("aut", "cre"),
+        comment = c(ORCID = "https://orcid.org/0000-0003-2172-5265")),
     person("Richard", "Ellison", role = "aut"),
     person("Bea", "Hernández", role = "rev",
         comment = "Bea reviewed the package for ropensci, see https://github.com/ropensci/onboarding/issues/116"),
@@ -47,3 +49,6 @@ BugReports: https://github.com/ropensci/bikedata/issues
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
+X-schema.org-applicationCategory: Data Access
+X-schema.org-keywords: bicycle-hire-systems, bike-hire-systems, bike-hire, bicycle-hire, database, bike-data
+X-schema.org-isPartOf: https://ropensci.org
