### 2017-12-04 a997874
version bump -> 0.1.0.99; add dodgr remote
