+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: bikedata
 Title: Download and Aggregate Data from Public Hire Bicycle Systems
-Version: 0.1.0
+Version: 0.1.0.099
 Authors@R: c(
     person("Mark", "Padgham", email = "mark.padgham@email.com",
         role = c("aut", "cre"),
@@ -41,6 +41,8 @@ Suggests:
 LinkingTo: 
     BH,
     Rcpp
+Remotes:
+    ATFutures/dodgr
 VignetteBuilder: knitr
 SystemRequirements: C++11
 NeedsCompilation: yes
